server/requirements.txt
fastapi
uvicorn
openai==0.27.6
python-dotenv==0.19.0
pydantic